+++ tests/NodaMoney.Tests/MoneyFormattableSpec/FormatWithCurrencySymbol.cs
@@ -17,11 +17,11 @@ public class FormatWithCurrencySymbol
     public void WhenCurrentCultureInvariant_ThenDecimalsFollowsCurrencyAndAmountFollowsCurrentCulture()
     {
         Thread.CurrentThread.CurrentCulture.Name.Should().Be(""); // Invariant Culture
-        _yen.ToString("C").Should().Be("(¥98,765)");
-        _euro.ToString("C").Should().Be("(€98,765.43)");
-        _dollar.ToString("C").Should().Be("($98,765.43)");
-        _dinar.ToString("C").Should().Be("(BD98,765.432)");
-        _swissFranc.ToString("C").Should().Be("(Fr.98,765.43)");
+        _yen.ToString("c").Should().Be("(¥98,765)");
+        _euro.ToString("c").Should().Be("(€98,765.43)");
+        _dollar.ToString("c").Should().Be("($98,765.43)");
+        _dinar.ToString("c").Should().Be("(BD98,765.432)");
+        _swissFranc.ToString("c").Should().Be("(Fr.98,765.43)");
     }
 
     [Fact]
@@ -31,11 +31,11 @@ public class FormatWithCurrencySymbol
         // The US default negative pattern for .NET4.8 is Pattern 0 `($n)`, where >.NET6.0 uses Pattern 1 `-$n`
 
         Thread.CurrentThread.CurrentCulture.Name.Should().Be("en-US");
-        _yen.ToString("C").Should().BeOneOf("-¥98,765", "(¥98,765)");
-        _euro.ToString("C").Should().BeOneOf("-€98,765.43", "(€98,765.43)");
-        _dollar.ToString("C").Should().BeOneOf("-$98,765.43", "($98,765.43)");
-        _dinar.ToString("C").Should().BeOneOf("-BD98,765.432", "(BD98,765.432)");
-        _swissFranc.ToString("C").Should().BeOneOf("-Fr.98,765.43", "(Fr.98,765.43)");
+        _yen.ToString("c").Should().BeOneOf("-¥98,765", "(¥98,765)");
+        _euro.ToString("c").Should().BeOneOf("-€98,765.43", "(€98,765.43)");
+        _dollar.ToString("c").Should().BeOneOf("-$98,765.43", "($98,765.43)");
+        _dinar.ToString("c").Should().BeOneOf("-BD98,765.432", "(BD98,765.432)");
+        _swissFranc.ToString("c").Should().BeOneOf("-Fr.98,765.43", "(Fr.98,765.43)");
     }
 
     [Fact]
@@ -43,11 +43,11 @@ public class FormatWithCurrencySymbol
     public void WhenCurrentCultureNL_ThenDecimalsFollowsCurrencyAndAmountFollowsCurrentCultureNL()
     {
         Thread.CurrentThread.CurrentCulture.Name.Should().Be("nl-NL");
-        _yen.ToString("C").Should().Be("¥ -98.765");
-        _euro.ToString("C").Should().Be("€ -98.765,43");
-        _dollar.ToString("C").Should().Be("$ -98.765,43");
-        _dinar.ToString("C").Should().Be("BD -98.765,432");
-        _swissFranc.ToString("C").Should().Be("Fr. -98.765,43");
+        _yen.ToString("c").Should().Be("¥ -98.765");
+        _euro.ToString("c").Should().Be("€ -98.765,43");
+        _dollar.ToString("c").Should().Be("$ -98.765,43");
+        _dinar.ToString("c").Should().Be("BD -98.765,432");
+        _swissFranc.ToString("c").Should().Be("Fr. -98.765,43");
     }
 
     [Fact]
@@ -55,11 +55,11 @@ public class FormatWithCurrencySymbol
     public void WhenCurrentCultureFR_ThenDecimalsFollowsCurrencyAndAmountFollowsCurrentCultureFR()
     {
         Thread.CurrentThread.CurrentCulture.Name.Should().Be("fr-FR");
-        _yen.ToString("C").Should().Be("-98 765 ¥");
-        _euro.ToString("C").Should().Be("-98 765,43 €");
-        _dollar.ToString("C").Should().Be("-98 765,43 $");
-        _dinar.ToString("C").Should().Be("-98 765,432 BD");
-        _swissFranc.ToString("C").Should().Be("-98 765,43 Fr.");
+        _yen.ToString("c").Should().Be("-98 765 ¥");
+        _euro.ToString("c").Should().Be("-98 765,43 €");
+        _dollar.ToString("c").Should().Be("-98 765,43 $");
+        _dinar.ToString("c").Should().Be("-98 765,432 BD");
+        _swissFranc.ToString("c").Should().Be("-98 765,43 Fr.");
     }
 
     [Fact]
@@ -67,11 +67,11 @@ public class FormatWithCurrencySymbol
     public void WhenZeroDecimals_ThenThisShouldSucceed()
     {
         Thread.CurrentThread.CurrentCulture.Name.Should().Be("nl-NL");
-        _yen.ToString("C0").Should().Be("¥ -98.765");
-        _euro.ToString("C0").Should().Be("€ -98.765");
-        _dollar.ToString("C0").Should().Be("$ -98.765");
-        _dinar.ToString("C0").Should().Be("BD -98.765");
-        _swissFranc.ToString("C0").Should().Be("Fr. -98.765");
+        _yen.ToString("c0").Should().Be("¥ -98.765");
+        _euro.ToString("c0").Should().Be("€ -98.765");
+        _dollar.ToString("c0").Should().Be("$ -98.765");
+        _dinar.ToString("c0").Should().Be("BD -98.765");
+        _swissFranc.ToString("c0").Should().Be("Fr. -98.765");
     }
 
     [Fact]
@@ -79,11 +79,11 @@ public class FormatWithCurrencySymbol
     public void WhenOneDecimals_ThenThisShouldSucceed()
     {
         Thread.CurrentThread.CurrentCulture.Name.Should().Be("nl-NL");
-        _yen.ToString("C1").Should().Be("¥ -98.765,0");
-        _euro.ToString("C1").Should().Be("€ -98.765,4");
-        _dollar.ToString("C1").Should().Be("$ -98.765,4");
-        _dinar.ToString("C1").Should().Be("BD -98.765,4");
-        _swissFranc.ToString("C1").Should().Be("Fr. -98.765,4");
+        _yen.ToString("c1").Should().Be("¥ -98.765,0");
+        _euro.ToString("c1").Should().Be("€ -98.765,4");
+        _dollar.ToString("c1").Should().Be("$ -98.765,4");
+        _dinar.ToString("c1").Should().Be("BD -98.765,4");
+        _swissFranc.ToString("c1").Should().Be("Fr. -98.765,4");
     }
 
     [Fact]
@@ -91,11 +91,11 @@ public class FormatWithCurrencySymbol
     public void WhenTwoDecimals_ThenThisShouldSucceed()
     {
         Thread.CurrentThread.CurrentCulture.Name.Should().Be("nl-NL");
-        _yen.ToString("C2").Should().Be("¥ -98.765,00");
-        _euro.ToString("C2").Should().Be("€ -98.765,43");
-        _dollar.ToString("C2").Should().Be("$ -98.765,43");
-        _dinar.ToString("C2").Should().Be("BD -98.765,43");
-        _swissFranc.ToString("C2").Should().Be("Fr. -98.765,43");
+        _yen.ToString("c2").Should().Be("¥ -98.765,00");
+        _euro.ToString("c2").Should().Be("€ -98.765,43");
+        _dollar.ToString("c2").Should().Be("$ -98.765,43");
+        _dinar.ToString("c2").Should().Be("BD -98.765,43");
+        _swissFranc.ToString("c2").Should().Be("Fr. -98.765,43");
     }
 
     [Fact]
@@ -103,11 +103,11 @@ public class FormatWithCurrencySymbol
     public void WhenThreeDecimals_ThenThisShouldSucceed()
     {
         Thread.CurrentThread.CurrentCulture.Name.Should().Be("nl-NL");
-        _yen.ToString("C3").Should().Be("¥ -98.765,000");
-        _euro.ToString("C3").Should().Be("€ -98.765,430");
-        _dollar.ToString("C3").Should().Be("$ -98.765,430");
-        _dinar.ToString("C3").Should().Be("BD -98.765,432");
-        _swissFranc.ToString("C3").Should().Be("Fr. -98.765,430");
+        _yen.ToString("c3").Should().Be("¥ -98.765,000");
+        _euro.ToString("c3").Should().Be("€ -98.765,430");
+        _dollar.ToString("c3").Should().Be("$ -98.765,430");
+        _dinar.ToString("c3").Should().Be("BD -98.765,432");
+        _swissFranc.ToString("c3").Should().Be("Fr. -98.765,430");
     }
 
     [Fact]
@@ -115,10 +115,10 @@ public class FormatWithCurrencySymbol
     public void WhenFourDecimals_ThenThisShouldSucceed()
     {
         Thread.CurrentThread.CurrentCulture.Name.Should().Be("nl-NL");
-        _yen.ToString("C4").Should().Be("¥ -98.765,0000");
-        _euro.ToString("C4").Should().Be("€ -98.765,4300");
-        _dollar.ToString("C4").Should().Be("$ -98.765,4300");
-        _dinar.ToString("C4").Should().Be("BD -98.765,4320");
-        _swissFranc.ToString("C4").Should().Be("Fr. -98.765,4300");
+        _yen.ToString("c4").Should().Be("¥ -98.765,0000");
+        _euro.ToString("c4").Should().Be("€ -98.765,4300");
+        _dollar.ToString("c4").Should().Be("$ -98.765,4300");
+        _dinar.ToString("c4").Should().Be("BD -98.765,4320");
+        _swissFranc.ToString("c4").Should().Be("Fr. -98.765,4300");
     }
 }