+++ CHANGELOG.md
@@ -7,10 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Next]
 
 ### Added
--
+- Format to Compact notation, like USD 1.2K, €3.4M (#111)
 
 ### Changed
-- ISO 4217 AMENDMENT NUMBER 180, Bulgaria will use EUR from 1 January 2026 and BGN will move to the historic list
+- ISO 4217 AMENDMENT NUMBER 180, Bulgaria will use EUR from 1 January 2026, and BGN will move to the historic list.
+- Formatting format 'C' has different behavior depending on capitalization. Format 'C' will use the international currency symbol (US$),
+  where 'c' will use the local currency code ($).
 
 ### Removed
 -